Feature separation in API project
