+++ templates/MicroserviceTemplate.Aspire/tests/MyService.Tests/ItemTests/ValidationTests/CreateItemValidationTests.cs
@@ -2,7 +2,7 @@ using MyService.Application.Item.Commands;
 using MyService.Application.Item.Dtos;
 using MyService.Application.Item.Validations;
 
-namespace MyService.Tests.ItemTests;
+namespace MyService.Tests.ItemTests.ValidationTests;
 
 public class CreateItemValidationTests
 {
